Launch more bad-ass instance for IBD, add downgrade instructions
Uses a 40 GB RAM drive and 10 GB dbcache to speed up initial
blockchain download. It then prunes the chain and copies the
data directory. The README tells user to downgrade and reboot.

docker-compose is launched as a service as of the first reboot.

Switch to Amazon Linux 2 for systemd support.