AWS: use Rails without Docker
Downgrade to Ruby 2.4 which is part of amazon-linux-extras. This
involves fewer dependencies than rvm.

Clone the repo again in /var/www/matreon to allow seperate upgrading
of the system overall (which might move to a new repo in the future)
and the Matreon Rails application specifically.

Use Nginx to serve static files and remove the need for rack-proxy
for Lightning Charge.

Removed all Docker references.

Fixed crontab.