Move some configuration to rails_helper.rb
Some code in rails_helper.rb is meant to be executed before the Rails
application is loaded. But spec_helper.rb, which is loaded before
rails_helper.rb, already loads the Rails application.

The original setup of the two files leaves loading of the Rails
application to rails_helper.rb.